refactor: separate constants file
